Background, animations, lvel parser and sound effects
