+++ Fall of the Forest/Assets/Prefabs/Ground/wooden box.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 8356766944688275067}
   - component: {fileID: 8356766944688275060}
   - component: {fileID: 8356766944688275061}
+  - component: {fileID: 8446197663921782758}
   m_Layer: 0
   m_Name: wooden box
   m_TagString: Untagged
@@ -95,3 +96,19 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!54 &8446197663921782758
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8356766944688275062}
+  serializedVersion: 2
+  m_Mass: 0.5
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_UseGravity: 1
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 8
+  m_CollisionDetection: 0
