+++ Fall of the Forest/Assets/Scripts/Parallaxing.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1edeb6ee2fb4684439ba51e783b4808e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
