--- Fall of the Forest/Assets/Scripts/SimpleSampleCharacterControl.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ba158ff05cf30224f94bfdddec8e7635
-timeCreated: 1486974282
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
