+++ Fall of the Forest/Assets/Scripts/pllax.cs
@@ -1,0 +1,31 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class pllax : MonoBehaviour
+{
+
+    private float length, startpos, startposY;
+    public GameObject cam;
+    public float parallaxEffect;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        startpos = transform.position.x;
+        startposY = transform.position.y;
+        length = GetComponent<SpriteRenderer>().bounds.size.x;
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        float temp = (cam.transform.position.x * (1 - parallaxEffect));
+        float dist = (cam.transform.position.x * parallaxEffect);
+
+        transform.position = new Vector3(startpos + dist, startposY, transform.position.z);
+
+        if (temp > startpos + length) startpos += length;
+        else if (temp < startpos - length) startpos -= length;
+    }
+}
