+++ Fall of the Forest/Assets/Scripts/pllax.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b22073ae15cbe44d99c5df53f87c806
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
