+++ Fall of the Forest/Assets/Sounds/Background/177830__trebblofang__horror-atmosphere-background.flac.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 3c0db7a276350c543b43a13842938078
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
