+++ Fall of the Forest/Assets/Sounds/Background/238251__steaq__dark-ambient-backgrounding-001.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 99330436d563ddb4a8bb4463dfb9b819
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
