+++ Fall of the Forest/Assets/Sounds/Background/515090__nick-simon-adams__piano-background-3.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 5f5c8046fee2f424eade421f0c9f718e
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
