+++ Fall of the Forest/Assets/Sounds/PlayerMovement/Launch_Jump.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 40582efef0b19af4c850be6e4d56d17a
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
